requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
celery==4.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-bootstrap-modal-forms==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
kombu==4.6.11
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
sentry-sdk==1.0.0
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.26.4
vine==1.3.0
waitress==2.0.0
whitenoise==5.2.0
